Declassify the random seed
This changes what is declassified and when.

Before, declassification only occurs during rejection sampling and what
is declassified is whether rejection occurs or not.

After, the full sampling of the matrix expects its seed to be public:
what is declassified is the public key when it is read from memory in
the IND-CPA encryption, when it is derived from the output of SHA3 in
keygen.